backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
anthropic==0.50.0
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-core==0.3.51
langgraph==0.3.21
httpx==0.28.1
sse-starlette==2.2.1
pydantic==2.11.2
python-multipart==0.0.20
